pyproject.toml
[build-system]
requires = ['flit_core >=3.4,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'aiida-3dd'
dynamic = ['version']
description = 'Package for the 3DD project.'
authors = [{name = 'Sebastiaan P. Huber', email = 'mail@sphuber.net'}]
readme = 'README.md'
license = {file = 'LICENSE.txt'}
classifiers = [
    'Development Status :: 4 - Development',
    'Framework :: AiiDA',
    'License :: OSI Approved :: MIT License',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
]
keywords = ['aiida']
requires-python = '>=3.8'
dependencies = [
    'aiida-core[atomic_tools]~=1.6',
    'click~=7.0',
    'click-completion',
    'tabulate',
]

[project.urls]
Home = 'https://github.com/sphuber/aiida-3dd'
Source = 'https://github.com/sphuber/aiida-3dd'

[project.optional-dependencies]
dev = [
    'pre-commit~=2.17',
    'pylint~=2.12.2',
    'pylint-aiida~=0.1',
    'toml'
]

[project.scripts]
aiida-3dd = 'aiida_3dd.cli:cmd_root'

[tool.flit.module]
name = 'aiida_3dd'

[tool.flynt]
line-length = 120
fail-on-change = true

[tool.isort]
force_sort_within_sections = true
include_trailing_comma = true
line_length = 120
multi_line_output = 3

[tool.pydocstyle]
ignore = [
    'D104',
    'D202',
    'D203',
    'D213',
    'D301',
]

[tool.pylint.basic]
good-names = ['pk']

[tool.pylint.master]
load-plugins = ['pylint_aiida']

[tool.pylint.format]
max-line-length = 120

[tool.pylint.messages_control]
disable = [
    'bad-continuation',
    'duplicate-code',
    'import-outside-toplevel',
    'too-many-arguments',
    'too-many-branches',
    'too-many-locals',
    'too-many-statements',
]

[tool.yapf]
align_closing_bracket_with_visual_indent = true
based_on_style = 'google'
coalesce_brackets = true
column_limit = 120
dedent_closing_brackets = true
indent_dictionary_value = false
split_arguments_when_comma_terminated = true
